Recap add update version of ResetDatabase /which delete and create database much faster/ and add Import Employees(add data anotation for some properties) for my exam from 10.12.2017 FastFood app.

+++ C#DBFundamentals/DB-Advanced-Entity-Framework-Core/MyExam10.12.2017/src/FastFood.App/Startup.cs
@@ -1,6 +1,8 @@
 ﻿using AutoMapper;
 using FastFood.Data;
+using Microsoft.EntityFrameworkCore;
 using System;
+using System.Data.SqlClient;
 using System.IO;
 
 namespace FastFood.App
@@ -17,19 +19,22 @@ namespace FastFood.App
 
             Mapper.Initialize(cfg => cfg.AddProfile<FastFoodProfile>());
 
-            //ImportEntities(context);
+            ImportEntities(context);
 
             //ExportEntities(context);
 
             //BonusTask(context);
         }
 
-		private static void ImportEntities(FastFoodDbContext context, string baseDir = @"..\Datasets\")
+		private static void ImportEntities(FastFoodDbContext context, string baseDir = @"C:\Users\tihom\source\SoftUniCoursesCSharp\00GitAndGitHub\SoftUni\C#DBFundamentals\DB-Advanced-Entity-Framework-Core\MyExam10.12.2017\src\Datasets\")
 		{
-			const string exportDir = "./ImportResults/";
+			const string exportDir = @"C:\Users\tihom\source\SoftUniCoursesCSharp\00GitAndGitHub\SoftUni\C#DBFundamentals\DB-Advanced-Entity-Framework-Core\MyExam10.12.2017\src\FastFood.App\ImportResults\";
 
-            var employees = DataProcessor.Deserializer.ImportEmployees(context, File.ReadAllText(baseDir + "employees.json"));
+
+			var employees = DataProcessor.Deserializer.ImportEmployees(context, File.ReadAllText(baseDir + "employees.json"));
             PrintAndExportEntityToFile(employees, exportDir + "Employees.txt");
+
+			return;
 
             var items = DataProcessor.Deserializer.ImportItems(context, File.ReadAllText(baseDir + "items.json"));
             PrintAndExportEntityToFile(items, exportDir + "Items.txt");
@@ -63,10 +68,33 @@ namespace FastFood.App
 			File.WriteAllText(outputPath, entityOutput.TrimEnd());
 		}
 
-		private static void ResetDatabase(FastFoodDbContext context)
+		private static void ResetDatabase(FastFoodDbContext context, bool shouldDeleteDatabase = false)
 		{
-			context.Database.EnsureDeleted();
+			if (shouldDeleteDatabase)
+			{
+				context.Database.EnsureDeleted();
+				context.Database.EnsureCreated();
+			}
+
 			context.Database.EnsureCreated();
+
+			var disableIntegrityChecksQuery = "EXEC sp_MSforeachtable @command1='ALTER TABLE ? NOCHECK CONSTRAINT ALL'";
+			context.Database.ExecuteSqlCommand(disableIntegrityChecksQuery);
+
+			var deleteRowsQuery = "EXEC sp_MSforeachtable @command1='DELETE FROM ?'";
+			context.Database.ExecuteSqlCommand(deleteRowsQuery);
+
+			var enableIntegrityChecksQuery = "EXEC sp_MSforeachtable @command1='ALTER TABLE ? WITH CHECK CHECK CONSTRAINT ALL'";
+			context.Database.ExecuteSqlCommand(enableIntegrityChecksQuery);
+
+			var reseedQuery = "EXEC sp_MSforeachtable @command1='DBCC CHECKIDENT(''?'', RESEED, 0)'";
+			try
+			{
+				context.Database.ExecuteSqlCommand(reseedQuery);
+			}
+			catch (SqlException) // OrderItems table has no identity column, which isn't a problem
+			{
+			}
 		}
 	}
 }
